Split askAI into askGemini and askOllama
- Add src/ai/askGemini.ts: full Gemini implementation with focused
  option types, GCS support, safety settings, web search, pricing
- Add src/ai/askOllama.ts: full Ollama implementation with focused
  option types, local-only files, context window, thinking flag
- Add src/ai/helpers/cache.ts: shared cache read/write utilities
- Add src/ai/helpers/processResponse.ts: shared JSON parse + clean + test
- Refactor src/ai/askAI.ts into a thin router (~6 lines) that delegates
  to askGemini or askOllama based on the ollama option / OLLAMA env var;
  all overloads and JSDoc preserved for backward compatibility
- Export askGemini, askOllama, GeminiDetailedResponse, OllamaDetailedResponse
  from src/index.ts alongside the existing exports
- Replace test/ai/askAI-Google.test.ts with test/ai/askGemini.test.ts
- Replace test/ai/askAI-Ollama.test.ts with test/ai/askOllama.test.ts
