+++ test/ai/askGemini.test.ts
@@ -1,6 +1,6 @@
 import "@std/dotenv/load";
 import { assertEquals, assertRejects } from "jsr:@std/assert";
-import askAI from "../../src/ai/askAI.ts";
+import askGemini from "../../src/ai/askGemini.ts";
 import { existsSync, rmSync } from "node:fs";
 import * as z from "zod";
 
@@ -11,7 +11,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   }
 
   Deno.test("should run the doc example", async () => {
-    const europeanCountries = await askAI(
+    const europeanCountries = await askGemini(
       `Give me a list of three countries in Northern Europe.`,
       {
         returnJson: true,
@@ -45,7 +45,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       verbose: true,
     });
     console.log(result);
@@ -54,7 +54,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with a detailed response", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       verbose: true,
       detailedResponse: true,
     });
@@ -70,19 +70,19 @@ if (typeof aiKey === "string" && aiKey !== "") {
       totalOutputTokens: 0,
       totalRequests: 0,
     };
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       verbose: false,
       metrics,
     });
     console.log(result);
     console.log(metrics);
-    const result2 = await askAI("What is the capital of Canada?", {
+    const result2 = await askGemini("What is the capital of Canada?", {
       verbose: false,
       metrics,
     });
     console.log(result2);
     console.log(metrics);
-    const result3 = await askAI("What is the capital of Spain?", {
+    const result3 = await askGemini("What is the capital of Spain?", {
       verbose: false,
       metrics,
     });
@@ -99,19 +99,19 @@ if (typeof aiKey === "string" && aiKey !== "") {
       totalOutputTokens: 0,
       totalRequests: 0,
     };
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       verbose: true,
       metrics,
     });
     console.log(result);
     console.log(metrics);
-    const result2 = await askAI("What is the capital of Canada?", {
+    const result2 = await askGemini("What is the capital of Canada?", {
       verbose: true,
       metrics,
     });
     console.log(result2);
     console.log(metrics);
-    const result3 = await askAI("What is the capital of Spain?", {
+    const result3 = await askGemini("What is the capital of Spain?", {
       verbose: true,
       metrics,
     });
@@ -122,7 +122,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with undefined thinking", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       verbose: true,
       model: "gemini-2.5-flash",
     });
@@ -132,7 +132,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt without thinking", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       thinkingBudget: 0,
       verbose: true,
       model: "gemini-2.5-flash",
@@ -143,7 +143,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with thinking (5000)", async () => {
-    const result = await askAI(
+    const result = await askGemini(
       "Find the sum of all integer bases b > 9 for which 17b is a divisor of 97b. Return just the result. No explanations. But think carefully first.",
       {
         thinkingBudget: 5000,
@@ -157,7 +157,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with dynamic thinking (-1) and detailed response and cache", async () => {
-    const result = await askAI(
+    const result = await askGemini(
       "Find the sum of all integer bases b > 9 for which 17b is a divisor of 97b. Return just the result. No explanations. But think carefully first.",
       {
         thinkingBudget: -1,
@@ -172,7 +172,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached response", async () => {
-    const result = await askAI(
+    const result = await askGemini(
       "Find the sum of all integer bases b > 9 for which 17b is a divisor of 97b. Return just the result. No explanations. But think carefully first.",
       {
         thinkingBudget: -1,
@@ -187,7 +187,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt without thinking and returning JSON", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       thinkingBudget: 0,
       verbose: true,
       returnJson: true,
@@ -199,7 +199,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with thinking and returning JSON", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       thinkingBudget: 500,
       verbose: true,
       returnJson: true,
@@ -211,7 +211,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with a test", async () => {
-    const result = await askAI("Give me a list of 3 countries in Europe.", {
+    const result = await askGemini("Give me a list of 3 countries in Europe.", {
       returnJson: true,
       test: (response: unknown) => {
         if (!Array.isArray(response)) {
@@ -234,7 +234,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with a list of tests", async () => {
-    const result = await askAI("Give me a list of 3 countries in Europe.", {
+    const result = await askGemini("Give me a list of 3 countries in Europe.", {
       returnJson: true,
       test: [(response: unknown) => {
         if (!Array.isArray(response)) {
@@ -257,7 +257,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       cache: true,
     });
     console.log(result);
@@ -266,7 +266,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached data", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       cache: true,
     });
     console.log(result);
@@ -275,7 +275,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and json", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       cache: true,
       returnJson: true,
     });
@@ -285,7 +285,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached JSON data", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       cache: true,
       returnJson: true,
     });
@@ -295,7 +295,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and verbose", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askGemini("What is the capital of Canada?", {
       cache: true,
       verbose: true,
     });
@@ -304,7 +304,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached data with verbose", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askGemini("What is the capital of Canada?", {
       cache: true,
       verbose: true,
     });
@@ -313,7 +313,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and verbose and json", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askGemini("What is the capital of Canada?", {
       cache: true,
       returnJson: true,
       verbose: true,
@@ -323,7 +323,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached json data with verbose and json", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askGemini("What is the capital of Canada?", {
       cache: true,
       returnJson: true,
       verbose: true,
@@ -334,7 +334,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should use a simple prompt and return a json", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askGemini("What is the capital of France?", {
       returnJson: true,
     });
     console.log(result);
@@ -344,7 +344,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should use a simple prompt and log extra information", async () => {
-    await askAI("What is the capital of France?", {
+    await askGemini("What is the capital of France?", {
       verbose: true,
     });
 
@@ -353,7 +353,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should scrape a web page", async () => {
-    await askAI(
+    await askGemini(
       `Here's the page showing presidential executive orders. Extract the executive order/names, dates (yyyy-mm-dd), and urls as an array of objects. Also categorize each executive order based on its name.`,
       {
         HTMLFrom:
@@ -368,7 +368,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should analyze an audio file", async () => {
-    const audioResponse = await askAI(
+    const audioResponse = await askGemini(
       `Return an object with the name of the person talking and an approximate date of the speech if you recognize it.`,
       {
         audio: "test/data/ai/speech.mp3",
@@ -388,7 +388,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
       images.push(`test/data/ai/pictures/${dirEntry.name}`);
     }
 
-    await askAI(
+    await askGemini(
       `Based on the images I send you, I want an array of objects with the following properties:
     - name: the person on the image if it's a human and you can recognize it,
     - description: a very short description of the image,
@@ -405,7 +405,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should analyze a video file", async () => {
-    await askAI(
+    await askGemini(
       `I want a array of objects, with each object having the following keys: name, timestamp, main emotion, transcript. Each time a new person talks, create a new object.`,
       {
         video: "test/data/ai/The Ontario leaders' debate in 3 minutes 360.mp4",
@@ -418,7 +418,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should analyze a pdf file with a specific model", async () => {
-    await askAI(
+    await askGemini(
       `This is a supreme court decision. Give me the merits of the case in the document. I want to know what happened and when. Return a list of objects with a date and a brief summary for each important event. Sort them chronologically.`,
       {
         model: "gemini-2.0-flash",
@@ -431,7 +431,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should analyze different files with different formats", async () => {
-    await askAI(
+    await askGemini(
       `Give me a short description of each things I give you.`,
       {
         model: "gemini-2.0-flash",
@@ -449,7 +449,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should return raw string when parseJson is false and returnJson is true", async () => {
-    const result = await askAI("Give me a list of 3 countries in Europe.", {
+    const result = await askGemini("Give me a list of 3 countries in Europe.", {
       returnJson: true,
       parseJson: false,
     });
@@ -464,7 +464,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
   });
 
   Deno.test("should use a text file", async () => {
-    await askAI(
+    await askGemini(
       "What is the content of this text file?",
       {
         text: "test/data/data.csv",
@@ -474,48 +474,48 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should use an image file stored in a google bucket", async () => {
-    await askAI("What is in this image?", {
+    await askGemini("What is in this image?", {
       verbose: true,
       image: "gs://nael_test_bucket/journalism-tests/cat.png",
     });
     assertEquals(true, true);
   });
   Deno.test("should use a video file stored in a google bucket", async () => {
-    await askAI("What is happening in this video?", {
+    await askGemini("What is happening in this video?", {
       verbose: true,
       video: "gs://nael_test_bucket/journalism-tests/debate.mp4",
     });
     assertEquals(true, true);
   });
   Deno.test("should use a pdf file stored in a google bucket", async () => {
-    await askAI("What is this document about?", {
+    await askGemini("What is this document about?", {
       verbose: true,
       pdf: "gs://nael_test_bucket/journalism-tests/piekut.pdf",
     });
     assertEquals(true, true);
   });
   Deno.test("should use an audio file stored in a google bucket", async () => {
-    await askAI("What is this audio about?", {
+    await askGemini("What is this audio about?", {
       verbose: true,
       audio: "gs://nael_test_bucket/journalism-tests/speech.mp3",
     });
     assertEquals(true, true);
   });
   Deno.test("should use a text file stored in a google bucket", async () => {
-    await askAI("What is the content of this text file?", {
+    await askGemini("What is the content of this text file?", {
       verbose: true,
       text: "gs://nael_test_bucket/journalism-tests/data.csv",
     });
     assertEquals(true, true);
   });
   Deno.test("should answer without grounding with web search", async () => {
-    await askAI("Who is Nael Shiab?", {
+    await askGemini("Who is Nael Shiab?", {
       verbose: true,
     });
     assertEquals(true, true);
   });
   Deno.test("should answer with grounding with web search and caching", async () => {
-    await askAI("Who is Nael Shiab?", {
+    await askGemini("Who is Nael Shiab?", {
       verbose: true,
       webSearch: true,
       cache: true,
@@ -523,7 +523,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should answer with grounding with web search and return cached data", async () => {
-    await askAI("Who is Nael Shiab?", {
+    await askGemini("Who is Nael Shiab?", {
       verbose: true,
       webSearch: true,
       cache: true,
@@ -540,7 +540,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
       })),
     );
 
-    await askAI("Give me 10 random people.", {
+    await askGemini("Give me 10 random people.", {
       verbose: true,
       cache: true,
       schemaJson: schema,
@@ -556,7 +556,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
       })),
     );
 
-    await askAI("Give me 10 random people.", {
+    await askGemini("Give me 10 random people.", {
       verbose: true,
       cache: true,
       schemaJson: schema,
@@ -564,7 +564,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should work with thinking level minimal by default", async () => {
-    await askAI("Give me 10 random people.", {
+    await askGemini("Give me 10 random people.", {
       verbose: true,
       cache: true,
       includeThoughts: true,
@@ -573,7 +573,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should work with thinking level medium", async () => {
-    await askAI("Give me 10 random people.", {
+    await askGemini("Give me 10 random people.", {
       verbose: true,
       cache: true,
       thinkingLevel: "medium",
@@ -583,20 +583,20 @@ if (typeof aiKey === "string" && aiKey !== "") {
     assertEquals(true, true);
   });
   Deno.test("should work without a system prompt", async () => {
-    await askAI("Why is the sky blue?", {
+    await askGemini("Why is the sky blue?", {
       verbose: true,
     });
     assertEquals(true, true);
   });
   Deno.test("should work with a system prompt", async () => {
-    await askAI("Why is the sky blue?", {
+    await askGemini("Why is the sky blue?", {
       verbose: true,
       systemPrompt: "Always answer with rhymes.",
     });
     assertEquals(true, true);
   });
   Deno.test("should accept safetyEnabled option", async () => {
-    await askAI("What is the capital of France?", {
+    await askGemini("What is the capital of France?", {
       safetyEnabled: false,
     });
     assertEquals(true, true);
@@ -608,7 +608,7 @@ if (typeof aiKey === "string" && aiKey !== "") {
 Deno.test("should throw an error when taking a screenshot", async () => {
   await assertRejects(
     () =>
-      askAI(
+      askGemini(
         `Tell me which products are on special.`,
         {
           screenshotFrom: "https://www.metro.ca/circulaire",
