+++ test/ai/askOllama.test.ts
@@ -1,18 +1,18 @@
 import "@std/dotenv/load";
 import { assertEquals, assertRejects } from "jsr:@std/assert";
-import askAI from "../../src/ai/askAI.ts";
+import askOllama from "../../src/ai/askOllama.ts";
 import { existsSync, rmSync } from "node:fs";
 import { Ollama } from "ollama";
 import * as z from "zod";
 
-const ollama = Deno.env.get("OLLAMA");
-console.log("OLLAMA", ollama);
-if (ollama) {
+const ollamaEnv = Deno.env.get("OLLAMA");
+console.log("OLLAMA", ollamaEnv);
+if (ollamaEnv) {
   if (existsSync("./.journalism-cache")) {
     rmSync("./.journalism-cache", { recursive: true });
   }
   Deno.test("should use a simple prompt (ollama)", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askOllama("What is the capital of France?", {
       verbose: true,
     });
     console.log(result);
@@ -21,7 +21,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with a high temperature (ollama)", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askOllama("What is the capital of France?", {
       verbose: true,
       temperature: 1,
     });
@@ -33,7 +33,7 @@ if (ollama) {
   Deno.test("should use a simple prompt with thinking (ollama)", {
     sanitizeResources: false,
   }, async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askOllama("What is the capital of France?", {
       verbose: true,
       thinkingBudget: 1,
     });
@@ -48,7 +48,7 @@ if (ollama) {
       sanitizeResources: false,
     },
     async () => {
-      const result = await askAI(
+      const result = await askOllama(
         "What is the capital of France? Return a JSON with this shape: {country: string, capital: string}",
         {
           verbose: true,
@@ -70,9 +70,11 @@ if (ollama) {
     "should use a simple prompt with a different Ollama instance (ollama)",
     { sanitizeResources: false },
     async () => {
-      const ollama = new Ollama({ host: "http://127.0.0.1:11434" });
+      const ollamaClient = new Ollama({ host: "http://127.0.0.1:11434" });
 
-      const result = await askAI("What is the capital of France?", { ollama });
+      const result = await askOllama("What is the capital of France?", {
+        ollama: ollamaClient,
+      });
       console.log(result);
 
       // Just making sure it doesn't crash for now.
@@ -80,7 +82,7 @@ if (ollama) {
     },
   );
   Deno.test("should use a simple prompt with a cleaning and test functions (ollama)", async () => {
-    const result = await askAI(
+    const result = await askOllama(
       "Give me a list of 3 countries in Europe.",
       {
         returnJson: true,
@@ -110,7 +112,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with a cleaning and test functions and return cached data (ollama)", async () => {
-    const result = await askAI(
+    const result = await askOllama(
       "Give me a list of 3 countries in Europe.",
       {
         returnJson: true,
@@ -142,7 +144,7 @@ if (ollama) {
   Deno.test("should use be able to clean complex response (ollama)", {
     sanitizeResources: false,
   }, async () => {
-    const result = await askAI(
+    const result = await askOllama(
       `Guess whether it's a "Man" or a "Woman". If it could be both, return "Neutral". Return an objects with two keys in it: one with the names as an array and the other with the genders as an array.
   Here are the name values as a JSON array:
   ["Marie","John","Alex"]
@@ -179,7 +181,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache (ollama)", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askOllama("What is the capital of France?", {
       cache: true,
     });
     console.log(result);
@@ -188,7 +190,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached data", async () => {
-    const result = await askAI("What is the capital of France?", {
+    const result = await askOllama("What is the capital of France?", {
       cache: true,
     });
     console.log(result);
@@ -197,27 +199,33 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and json", async () => {
-    const result = await askAI("What is the capital of France? Return a JSON", {
-      cache: true,
-      returnJson: true,
-    });
+    const result = await askOllama(
+      "What is the capital of France? Return a JSON",
+      {
+        cache: true,
+        returnJson: true,
+      },
+    );
     console.log(result);
 
     // Just making sure it doesn't crash for now.
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached JSON data (ollama)", async () => {
-    const result = await askAI("What is the capital of France? Return a JSON", {
-      cache: true,
-      returnJson: true,
-    });
+    const result = await askOllama(
+      "What is the capital of France? Return a JSON",
+      {
+        cache: true,
+        returnJson: true,
+      },
+    );
     console.log(result);
 
     // Just making sure it doesn't crash for now.
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and verbose", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askOllama("What is the capital of Canada?", {
       cache: true,
       verbose: true,
     });
@@ -226,7 +234,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached data with verbose (ollama)", async () => {
-    await askAI("What is the capital of Canada?", {
+    await askOllama("What is the capital of Canada?", {
       cache: true,
       verbose: true,
     });
@@ -235,7 +243,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt with cache and verbose and json (ollama)", async () => {
-    await askAI("What is the capital of Canada? Return a JSON.", {
+    await askOllama("What is the capital of Canada? Return a JSON.", {
       cache: true,
       returnJson: true,
       verbose: true,
@@ -245,7 +253,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and return cached json data with verbose and json (ollama)", async () => {
-    await askAI("What is the capital of Canada? Return a JSON.", {
+    await askOllama("What is the capital of Canada? Return a JSON.", {
       cache: true,
       returnJson: true,
       verbose: true,
@@ -255,7 +263,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should use a simple prompt and log extra information (ollama)", async () => {
-    await askAI("What is the capital of France?", {
+    await askOllama("What is the capital of France?", {
       verbose: true,
     });
 
@@ -267,7 +275,7 @@ if (ollama) {
     "should scrape a web page with default context window size (ollama)",
     { sanitizeResources: false },
     async () => {
-      await askAI(
+      await askOllama(
         `What is this website about?`,
         {
           HTMLFrom: "https://www.code-like-a-journalist.com/en",
@@ -286,7 +294,7 @@ if (ollama) {
     "should scrape a web page with specific context window size (ollama)",
     { sanitizeResources: false },
     async () => {
-      await askAI(
+      await askOllama(
         `What is this website about?`,
         {
           HTMLFrom: "https://www.code-like-a-journalist.com/en",
@@ -305,7 +313,7 @@ if (ollama) {
     "should analyze images (ollama)",
     { sanitizeResources: false },
     async () => {
-      await askAI(
+      await askOllama(
         `I want an object with the following properties:
         - name: the person on the image,
         - description: a very short description of the image,
@@ -324,7 +332,7 @@ if (ollama) {
     },
   );
   Deno.test("should use a text file (ollama)", async () => {
-    const result = await askAI(
+    const result = await askOllama(
       "What is the content of this text file?",
       {
         text: "test/data/data.csv",
@@ -335,7 +343,7 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should return raw string when parseJson is false and returnJson is true", async () => {
-    const result = await askAI("Give me a list of 3 countries in Europe.", {
+    const result = await askOllama("Give me a list of 3 countries in Europe.", {
       returnJson: true,
       parseJson: false,
     });
@@ -361,7 +369,7 @@ if (ollama) {
       }),
     );
 
-    await askAI("Give me 10 random people.", {
+    await askOllama("Give me 10 random people.", {
       verbose: true,
       cache: true,
       schemaJson: schema,
@@ -381,7 +389,7 @@ if (ollama) {
       }),
     );
 
-    await askAI("Give me 10 random people.", {
+    await askOllama("Give me 10 random people.", {
       verbose: true,
       cache: true,
       schemaJson: schema,
@@ -389,13 +397,13 @@ if (ollama) {
     assertEquals(true, true);
   });
   Deno.test("should work without a system prompt", async () => {
-    await askAI("Why is the sky blue?", {
+    await askOllama("Why is the sky blue?", {
       verbose: true,
     });
     assertEquals(true, true);
   });
   Deno.test("should work with a system prompt", async () => {
-    await askAI("Why is the sky blue?", {
+    await askOllama("Why is the sky blue?", {
       verbose: true,
       systemPrompt: "Always answer with rhymes.",
     });
@@ -404,7 +412,7 @@ if (ollama) {
   Deno.test("should work with thinking level low by default", {
     sanitizeResources: false,
   }, async () => {
-    await askAI(
+    await askOllama(
       "Give me 10 random people (name, age, nationality, gender, profession). Make sure they are diverse.",
       {
         verbose: true,
@@ -417,7 +425,7 @@ if (ollama) {
   Deno.test("should work with thinking level low", {
     sanitizeResources: false,
   }, async () => {
-    await askAI(
+    await askOllama(
       "Give me 10 random people (name, age, nationality, gender, profession). Make sure they are diverse.",
       {
         verbose: true,
@@ -430,7 +438,7 @@ if (ollama) {
   Deno.test("should work with thinking level medium", {
     sanitizeResources: false,
   }, async () => {
-    await askAI(
+    await askOllama(
       "Give me 10 random people (name, age, nationality, gender, profession). Make sure they are diverse.",
       {
         verbose: true,
@@ -443,7 +451,7 @@ if (ollama) {
   Deno.test("should work with thinking level high", {
     sanitizeResources: false,
   }, async () => {
-    await askAI(
+    await askOllama(
       "Give me 10 random people (name, age, nationality, gender, profession). Make sure they are diverse.",
       {
         verbose: true,
@@ -457,19 +465,14 @@ if (ollama) {
   console.log("No OLLAMA in process.env");
 }
 
-Deno.test("should throw an error when taking a screenshot (ollama)", {
-  sanitizeResources: false,
-}, async () => {
+Deno.test("should throw an error for GCS files (ollama)", async () => {
   await assertRejects(
     () =>
-      askAI(
-        `Tell me which products are on special.`,
-        {
-          screenshotFrom: "https://www.metro.ca/circulaire",
-          verbose: true,
-        },
+      askOllama(
+        `What is in this file?`,
+        { text: "gs://some-bucket/file.csv" },
       ),
     Error,
-    "The 'screenshotFrom' option has been removed to reduce dependencies. Please take a screenshot yourself and pass it via the 'image' option.",
+    "Ollama does not support Google Cloud Storage files.",
   );
 });
